+++ R/stat-dens2d-labels.r
@@ -41,9 +41,13 @@
 #'
 #'   Computation of density and of the default bandwidth require at least
 #'   two observations with different values. If data do not fulfill this
-#'   condition, they are kept only if `keep.fraction = 1`. This is correct
+#'   condition, they are kept only if \code{keep.fraction = 1}. This is correct
 #'   behavior for a single observation, but can be surprising in the case of
 #'   multiple observations.
+#'
+#'   Parameters \code{keep.these} and \code{exclude.these} make it possible to
+#'   force inclusion or exclusion of observations after the density is computed.
+#'   In case of conflict, \code{exclude.these} overrides \code{keep.these}.
 #'
 #' @note Which points are kept and which not depends on how dense a grid is used
 #'   and how flexible the density surface estimate is. This depends on the
@@ -67,15 +71,15 @@
 #' @param keep.sparse logical If \code{TRUE}, the default, observations from the
 #'   more sparse regions are retained, if \code{FALSE} those from the densest
 #'   regions.
-#' @param keep.these character vector, integer vector, logical vector or
-#'   function that takes the variable mapped to the \code{label} aesthetic as
-#'   first argument and returns a character vector or a logical vector. Negative
-#'   integers behave as in R's extraction methods. The rows from \code{data}
-#'   indicated by \code{keep.these} are kept irrespective of the local density.
-#' @param keep.these.target character or \code{NULL}. Name of the column of
-#'   \code{data}, corresponding to a mapped aesthetic or group, to pass as first
-#'   argument to the function passed as argument to \code{keep.these}. If
-#'   \code{NULL} the whole \code{data} object is passed.
+#' @param keep.these,exclude.these character vector, integer vector, logical
+#'   vector or function that takes one or more variables in data selected by
+#'   \code{these.target}. Negative integers behave as in R's extraction methods.
+#'   The rows from \code{data} indicated by \code{keep.these} and
+#'   \code{exclude.these} are kept or excluded irrespective of the local
+#'   density.
+#' @param these.target character, numeric or logical selecting one or more
+#'   column(s) of \code{data}. If \code{TRUE} the whole \code{data} object is
+#'   passed.
 #' @param pool.along character, one of \code{"none"} or \code{"x"},
 #'   indicating if selection should be done pooling the observations along the
 #'   \emph{x} aesthetic, or separately on either side of \code{xintercept}.
@@ -218,7 +222,8 @@ stat_dens2d_labels <-
            keep.number = Inf,
            keep.sparse = TRUE,
            keep.these = FALSE,
-           keep.these.target = "label",
+           exclude.these = FALSE,
+           these.target = "label",
            pool.along = "xy",
            xintercept = 0,
            yintercept = 0,
@@ -267,7 +272,8 @@ stat_dens2d_labels <-
                     keep.number = keep.number,
                     keep.sparse = keep.sparse,
                     keep.these = keep.these,
-                    keep.these.target = keep.these.target,
+                    exclude.these = exclude.these,
+                    these.target = these.target,
                     pool.along = pool.along,
                     xintercept = xintercept,
                     yintercept = yintercept,
@@ -296,7 +302,8 @@ StatDens2dLabels <-
                keep.number,
                keep.sparse,
                keep.these,
-               keep.these.target,
+               exclude.these,
+               these.target,
                pool.along,
                xintercept,
                yintercept,
@@ -311,9 +318,13 @@ StatDens2dLabels <-
           data[["label"]] <- rownames(data)
         }
 
-        keep.these <- keep_these2logical(keep.these = keep.these,
-                                         data = data,
-                                         keep.these.target = keep.these.target)
+        keep.these <- these2logical(these = keep.these,
+                                    data = data,
+                                    these.target = these.target)
+
+        exclude.these <- these2logical(these = exclude.these,
+                                       data = data,
+                                       these.target = these.target)
 
         # discard redundant splits
         if (pool.along != "xy") {
@@ -440,6 +451,7 @@ StatDens2dLabels <-
             }
           }
         }
+        keep <- keep & !exclude.these
 
         if (invert.selection){
           keep <- !keep
@@ -487,50 +499,50 @@ StatDens2dLabels <-
 #'
 #' @keywords internal
 #'
-keep_these2logical <- function(keep.these,
-                               data,
-                               keep.these.target = "label") {
-  if (length(keep.these)) {
-    if (is.character(keep.these) || is.function(keep.these)) {
-      if (is.character(keep.these.target) && any(keep.these.target == "label") &&
+these2logical<- function(these,
+                         data,
+                         these.target = "label") {
+  if (length(these)) {
+    if (is.character(these) || is.function(these)) {
+      if (is.character(these.target) && any(these.target == "label") &&
           !exists("label", where = data, mode = "character", inherits = FALSE)) {
         data$label <- rownames(data)
       }
-      if (is.character(keep.these.target)) {
-        orig.num.targets <- length(unique(keep.these.target))
-        keep.these.target <- intersect(keep.these.target, colnames(data))
-        if (length(keep.these.target) == 0L) {
-          stop("Variables in 'keep.these.target' not in 'data'")
-        } else if (orig.num.targets > length(keep.these.target)) {
-          warning("Some variables in 'keep.these.target' not in 'data'")
+      if (is.character(these.target)) {
+        orig.num.targets <- length(unique(these.target))
+        these.target <- intersect(these.target, colnames(data))
+        if (length(these.target) == 0L) {
+          stop("Variables in 'these.target' not in 'data'")
+        } else if (orig.num.targets > length(these.target)) {
+          warning("Some variables in 'these.target' not in 'data'")
         }
       }
     }
-    if (is.function(keep.these)) {
-      keep.these <- keep.these(data[ , keep.these.target, drop = TRUE]) # any vector
+    if (is.function(these)) {
+      these <- these(data[ , these.target, drop = TRUE]) # any vector
     }
-    if (is.character(keep.these)) {
-      stopifnot(is.character(data[[keep.these.target[1]]]))
-      keep.these <- data[[keep.these.target[1]]] %in% keep.these # logical vector
+    if (is.character(these)) {
+      stopifnot(is.character(data[[these.target[1]]]))
+      these <- data[[these.target[1]]] %in% these # logical vector
     }
-    if (is.numeric(keep.these)) { # positional indices
+    if (is.numeric(these)) { # positional indices
       temp <- rep_len(FALSE, length.out = nrow(data))
-      temp[keep.these] <- TRUE
-      keep.these <- temp
+      temp[these] <- TRUE
+      these <- temp
     }
-    if (is.logical(keep.these)) { # logical indices, if short recycle
-      if (length(keep.these) >= 1L && length(keep.these) < nrow(data)) {
-        keep.these <- rep_len(keep.these, length.out = nrow(data))
-      } else if (length(keep.these) > nrow(data)) {
-        stop("Logical vector 'keep.these' longer than data")
+    if (is.logical(these)) { # logical indices, if short recycle
+      if (length(these) >= 1L && length(these) < nrow(data)) {
+        these <- rep_len(these, length.out = nrow(data))
+      } else if (length(these) > nrow(data)) {
+        stop("Logical vector 'keep.these' or 'exclude.these' longer than data")
       }
     }
-    if (anyNA(keep.these)) {
-      warning("Discarding 'NA's in keep.these")
-      keep.these[is.na(keep.these)] <- FALSE
+    if (anyNA(these)) {
+      warning("Discarding 'NA's in 'keep.these' or 'exclude-these'")
+      these[is.na(these)] <- FALSE
     }
   } else { # replace NULL and vectors with length zero with FALSE
-    keep.these <- rep(FALSE, nrow(data))
+    these <- rep_len(FALSE, length.out = nrow(data))
   }
-  keep.these
+  these
 }