+++ tests/testthat/test-stat-d2d-flt.R
@@ -733,54 +733,54 @@ test_that("labels_tb", {
   )
 })
 
-test_that("keep_these2logical arguments with label", {
+test_that("these2logicalarguments with label", {
   df <- data.frame(x = 1:10, y = 1:10, label = letters[1:10])
-  expect_equal(keep_these2logical(FALSE, df), rep_len(FALSE, nrow(df)))
-  expect_warning(keep_these2logical(NA, df))
-  expect_equal(keep_these2logical(NULL, df), rep_len(FALSE, nrow(df)))
-  expect_equal(keep_these2logical(numeric(), df), rep_len(FALSE, nrow(df)))
-  expect_error(keep_these2logical(nrow(df) + 1, df))
-  expect_equal(keep_these2logical(logical(), df), rep_len(FALSE, nrow(df)))
-  expect_equal(keep_these2logical(3:4, df), c(F, F, T, T, F, F, F, F, F, F))
-  expect_equal(keep_these2logical(-c(1:2, 5:10), df), c(F, F, T, T, F, F, F, F, F, F))
-  expect_equal(keep_these2logical(c("c", "d"), df), c(F, F, T, T, F, F, F, F, F, F))
-  expect_equal(keep_these2logical(function(x) {x %in% c("c", "d")}, df), c(F, F, T, T, F, F, F, F, F, F))
-  expect_equal(keep_these2logical(c("3", "4"), df), rep_len(FALSE, nrow(df)))
-  expect_equal(keep_these2logical(function(x) {x %in% c("3", "4")}, df), rep_len(FALSE, nrow(df)))
-  expect_equal(keep_these2logical(function(x) {x %in% c(3, 4)}, df,
-                                  keep.these.target = "x"), c(F, F, T, T, F, F, F, F, F, F))
-  expect_equal(keep_these2logical(function(d) {d$x %in% c(3, 4)}, df,
-                                  keep.these.target = c("x", "y")), c(F, F, T, T, F, F, F, F, F, F))
-  expect_equal(keep_these2logical(function(d) {d$label %in% c("c", "d")}, df,
-                                  keep.these.target = c("x", "y", "label")), c(F, F, T, T, F, F, F, F, F, F))
-  expect_equal(keep_these2logical(function(d) {d$x %in% c(3, 4)}, df,
-                                  keep.these.target = TRUE), c(F, F, T, T, F, F, F, F, F, F))
-  expect_error(keep_these2logical(function(d) {d$x %in% c(3, 4)}, df,
-                                  keep.these.target = "zz"))
-  expect_warning(keep_these2logical(function(x) {x %in% c(3, 4)}, df,
-                                  keep.these.target = c("zz", "x")))
+  expect_equal(these2logical(FALSE, df), rep_len(FALSE, nrow(df)))
+  expect_warning(these2logical(NA, df))
+  expect_equal(these2logical(NULL, df), rep_len(FALSE, nrow(df)))
+  expect_equal(these2logical(numeric(), df), rep_len(FALSE, nrow(df)))
+  expect_error(these2logical(nrow(df) + 1, df))
+  expect_equal(these2logical(logical(), df), rep_len(FALSE, nrow(df)))
+  expect_equal(these2logical(3:4, df), c(F, F, T, T, F, F, F, F, F, F))
+  expect_equal(these2logical(-c(1:2, 5:10), df), c(F, F, T, T, F, F, F, F, F, F))
+  expect_equal(these2logical(c("c", "d"), df), c(F, F, T, T, F, F, F, F, F, F))
+  expect_equal(these2logical(function(x) {x %in% c("c", "d")}, df), c(F, F, T, T, F, F, F, F, F, F))
+  expect_equal(these2logical(c("3", "4"), df), rep_len(FALSE, nrow(df)))
+  expect_equal(these2logical(function(x) {x %in% c("3", "4")}, df), rep_len(FALSE, nrow(df)))
+  expect_equal(these2logical(function(x) {x %in% c(3, 4)}, df,
+                                  these.target = "x"), c(F, F, T, T, F, F, F, F, F, F))
+  expect_equal(these2logical(function(d) {d$x %in% c(3, 4)}, df,
+                                  these.target = c("x", "y")), c(F, F, T, T, F, F, F, F, F, F))
+  expect_equal(these2logical(function(d) {d$label %in% c("c", "d")}, df,
+                                  these.target = c("x", "y", "label")), c(F, F, T, T, F, F, F, F, F, F))
+  expect_equal(these2logical(function(d) {d$x %in% c(3, 4)}, df,
+                                  these.target = TRUE), c(F, F, T, T, F, F, F, F, F, F))
+  expect_error(these2logical(function(d) {d$x %in% c(3, 4)}, df,
+                                  these.target = "zz"))
+  expect_warning(these2logical(function(x) {x %in% c(3, 4)}, df,
+                                  these.target = c("zz", "x")))
 })
 
-test_that("keep_these2logical arguments without label", {
+test_that("these2logicalarguments without label", {
   df <- data.frame(x = 1:10, y = 1:10)
-  expect_equal(keep_these2logical(FALSE, df), rep_len(FALSE, nrow(df)))
-  expect_warning(keep_these2logical(NA, df))
-  expect_equal(keep_these2logical(NULL, df), rep_len(FALSE, nrow(df)))
-  expect_equal(keep_these2logical(numeric(), df), rep_len(FALSE, nrow(df)))
-  expect_error(keep_these2logical(nrow(df) + 1, df))
-  expect_equal(keep_these2logical(logical(), df), rep_len(FALSE, nrow(df)))
-  expect_equal(keep_these2logical(3:4, df), c(F, F, T, T, F, F, F, F, F, F))
-  expect_equal(keep_these2logical(-c(1:2, 5:10), df), c(F, F, T, T, F, F, F, F, F, F))
-  expect_equal(keep_these2logical(c("c", "d"), df), rep_len(FALSE, nrow(df)))
-  expect_equal(keep_these2logical(function(x) {x %in% c("c", "d")}, df), rep_len(FALSE, nrow(df)))
-  expect_equal(keep_these2logical(c("3", "4"), df), c(F, F, T, T, F, F, F, F, F, F))
-  expect_equal(keep_these2logical(function(x) {x %in% c("3", "4")}, df), c(F, F, T, T, F, F, F, F, F, F))
-  expect_equal(keep_these2logical(function(x) {x %in% c(3, 4)}, df,
-                                  keep.these.target = "x"), c(F, F, T, T, F, F, F, F, F, F))
-  expect_equal(keep_these2logical(function(d) {d$x %in% c(3, 4)}, df,
-                                  keep.these.target = c("x", "y")), c(F, F, T, T, F, F, F, F, F, F))
-  expect_equal(keep_these2logical(function(d) {d$label %in% c("3", "4")}, df,
-                                  keep.these.target = c("x", "y", "label")), c(F, F, T, T, F, F, F, F, F, F))
-  expect_equal(keep_these2logical(function(d) {d$x %in% c(3, 4)}, df,
-                                  keep.these.target = TRUE), c(F, F, T, T, F, F, F, F, F, F))
+  expect_equal(these2logical(FALSE, df), rep_len(FALSE, nrow(df)))
+  expect_warning(these2logical(NA, df))
+  expect_equal(these2logical(NULL, df), rep_len(FALSE, nrow(df)))
+  expect_equal(these2logical(numeric(), df), rep_len(FALSE, nrow(df)))
+  expect_error(these2logical(nrow(df) + 1, df))
+  expect_equal(these2logical(logical(), df), rep_len(FALSE, nrow(df)))
+  expect_equal(these2logical(3:4, df), c(F, F, T, T, F, F, F, F, F, F))
+  expect_equal(these2logical(-c(1:2, 5:10), df), c(F, F, T, T, F, F, F, F, F, F))
+  expect_equal(these2logical(c("c", "d"), df), rep_len(FALSE, nrow(df)))
+  expect_equal(these2logical(function(x) {x %in% c("c", "d")}, df), rep_len(FALSE, nrow(df)))
+  expect_equal(these2logical(c("3", "4"), df), c(F, F, T, T, F, F, F, F, F, F))
+  expect_equal(these2logical(function(x) {x %in% c("3", "4")}, df), c(F, F, T, T, F, F, F, F, F, F))
+  expect_equal(these2logical(function(x) {x %in% c(3, 4)}, df,
+                                  these.target = "x"), c(F, F, T, T, F, F, F, F, F, F))
+  expect_equal(these2logical(function(d) {d$x %in% c(3, 4)}, df,
+                                  these.target = c("x", "y")), c(F, F, T, T, F, F, F, F, F, F))
+  expect_equal(these2logical(function(d) {d$label %in% c("3", "4")}, df,
+                                  these.target = c("x", "y", "label")), c(F, F, T, T, F, F, F, F, F, F))
+  expect_equal(these2logical(function(d) {d$x %in% c(3, 4)}, df,
+                                  these.target = TRUE), c(F, F, T, T, F, F, F, F, F, F))
 })