+++ NEWS.md
@@ -17,10 +17,14 @@ estimated (Issue #32 raised by mshubert to report bug).
 `stat_dens1d_filter_g()` use `type = 8` when computing quantiles, instead of the
 default `type = 7`. Type 8 is recommended as better with data that are not
 normally distributed.
+-   Add parameter `exclude.these` to `stat_dens2d_labels()`, `stat_dens2d_filter()`,
+`stat_dens2d_filter_g()`, `stat_dens1d_labels()`, `stat_dens1d_filter()` and
+`stat_dens1d_filter_g()` allowing forcing the exclusion of specific observations
+irrespective of the local density.
 -   In `stat_dens2d_labels()`, `stat_dens2d_filter()`,
 `stat_dens2d_filter_g()`, `stat_dens1d_labels()`, `stat_dens1d_filter()` and
-`stat_dens1d_filter_g()` add parameter `keep.these.target` making it possible
-to apply a function or vector of values to any variable in the layer `data` 
+`stat_dens1d_filter_g()` add parameter `these.target` making it possible
+to apply a function or vector of values passed through `exclude.these` or `keep.these` to any variable in the layer `data` 
 (See issue #19, partly addressing a suggestion by mshubert).
 -   Add unit tests and make existing tests visible to package 'covr'. (See
 issue #10, and pull requests #12, #13, #17, #18, #21, #23, #24, #25, #26, #27,